perf(web): prerender the charging map and cache its readers
